requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
aioredis==1.3.1
async-timeout==3.0.1
attrs==19.3.0
cachetools==4.1.1
cchardet==2.1.6
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
cytoolz==0.10.1
dataclasses==0.6
ecdsa==0.15
envparse==0.2.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-typing==2.2.1
eth-utils==1.9.0
expiringdict==1.2.1
grpcio==1.30.0
hiredis==1.0.1
idna==2.10
lmdb==0.98
more-itertools==8.4.0
multidict==4.7.6
protobuf==3.12.2
pycares==3.1.1
pycparser==2.20
pycryptodome==3.9.8
pyrchain==0.3.7
python-bitcoinlib==0.11.0
PyYAML==5.3.1
six==1.15.0
toolz==0.10.0
yarl==1.4.2
